third_party/WebKit/LayoutTests/fast/css/pseudo-required-optional-005-expected.txt

This test checks the CSS pseudoclasses for input types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isOptional("range") is true
PASS isOptional("submit") is true
PASS isOptional("image") is true
PASS isOptional("reset") is true
PASS isOptional("button") is true
PASS successfullyParsed is true

TEST COMPLETE

    
